% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{calculateTotalLipidScoresTableData}
\alias{calculateTotalLipidScoresTableData}
\title{Calculate the total lipid scores table data.}
\usage{
calculateTotalLipidScoresTableData(lipidScoresTableData)
}
\arguments{
\item{lipidScoresTableData}{The lipid scores table data.}
}
\value{
A tibble with the total lipid scores table data.
}
\description{
Calculate the total lipid scores table data.
}
