% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{checkNames}
\alias{checkNames}
\title{Check if the lipid names are valid shorthand names using the rgoslin package.}
\usage{
checkNames(totalLipidScoresTableData, cvMapTable)
}
\arguments{
\item{totalLipidScoresTableData}{The total lipid scores table data.}

\item{cvMapTable}{The shorthand CV map table.}
}
\value{
A tibble with the total lipid scores table data, joined with rgoslin parsing results.
}
\description{
Check if the lipid names are valid shorthand names using the rgoslin package.
}
