% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol-package.R
\docType{package}
\name{eposmol}
\alias{eposmol-package}
\alias{eposmol}
\title{Eposmol}
\description{
Eposmol is a package to score lipidomics data based on an additive, empirical scoring scheme.
}
\details{
The package provides a set of functions to read and score lipidomics identification data based on a scoring scheme. The package also provides a set of functions to parse lipid names and check if they are valid shorthand names using the rgoslin package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lifs-tools/empirical-lipid-ms-score}
  \item Report bugs at \url{https://github.com/lifs-tools/empirical-lipid-ms-score/issues}
}

}
\author{
\strong{Maintainer}: Nils Hoffmann \email{n.hoffmann@fz-juelich.de} (\href{https://orcid.org/0000-0002-6540-6875}{ORCID})

Authors:
\itemize{
  \item Harald Köfeler \email{harald.koefeler@medunigraz.at} (\href{https://orcid.org/0000-0002-2725-9616}{ORCID})
}

}
\keyword{internal}
