% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{loadCategoryAndClassMapTable}
\alias{loadCategoryAndClassMapTable}
\title{Load the category and class map table.}
\usage{
loadCategoryAndClassMapTable(
  path = system.file("..", "inst", "extdata", "class_map.xlsx")
)
}
\arguments{
\item{path}{The path to the category and class map table.}
}
\value{
A tibble with the category and class map table.
}
\description{
Load the category and class map table.
}
