% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{loadCvMapTable}
\alias{loadCvMapTable}
\title{Load the shorthand CV map table.}
\usage{
loadCvMapTable(
  path = system.file("..", "inst", "extdata", "shorthand_cv_map.xlsx")
)
}
\arguments{
\item{path}{The path to the shorthand CV map table.}
}
\value{
A tibble with the shorthand CV map table.
}
\description{
Load the shorthand CV map table.
}
