% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{loadScoringTable}
\alias{loadScoringTable}
\title{Load the scoring scheme table.}
\usage{
loadScoringTable(path = system.file("..", "inst", "extdata", "Table 1.xlsx"))
}
\arguments{
\item{path}{The path to the scoring scheme table.}
}
\value{
A tibble with the scoring scheme table.
}
\description{
Load the scoring scheme table.
}
