% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{readLongTable}
\alias{readLongTable}
\title{Read a long format table and join with the features defined in the scoring table. The long format table must have the following columns: Name, LipidCategoryOrClass, IonMode, Feature, Value.}
\usage{
readLongTable(tble, scoringTable)
}
\arguments{
\item{tble}{The long format table.}

\item{scoringTable}{The scoring table.}
}
\value{
A tibble with the long format table.
}
\description{
Read a long format table and join with the features defined in the scoring table. The long format table must have the following columns: Name, LipidCategoryOrClass, IonMode, Feature, Value.
}
