% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{readManualTable}
\alias{readManualTable}
\title{Read the manual table and join with the features defined in the scoring table. The manual table must have the following columns: Name, LipidCategoryOrClass, IonMode, Feature, Value.}
\usage{
readManualTable(tble, scoringTable)
}
\arguments{
\item{tble}{The manual table.}

\item{scoringTable}{The scoring table.}
}
\value{
A tibble with the manual table data.
}
\description{
Read the manual table and join with the features defined in the scoring table. The manual table must have the following columns: Name, LipidCategoryOrClass, IonMode, Feature, Value.
}
