% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eposmol.R
\name{readWideTable}
\alias{readWideTable}
\title{Read a wide format table and join with the features defined in the scoring table. The wide format table must have the following columns: Name, LipidCategoryOrClass, IonMode, and the features as columns.}
\usage{
readWideTable(tble, scoringTable)
}
\arguments{
\item{tble}{The wide format table.}

\item{scoringTable}{The scoring table.}
}
\value{
A tibble with the wide format table.
}
\description{
Read a wide format table and join with the features defined in the scoring table. The wide format table must have the following columns: Name, LipidCategoryOrClass, IonMode, and the features as columns.
}
