% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_eposmol_app}
\alias{run_eposmol_app}
\title{Start the EPoSMoL Shiny Webapplication.}
\usage{
run_eposmol_app(test.mode = FALSE)
}
\arguments{
\item{test.mode}{enable or disable test.mode}
}
\value{
A ShinyApp object.
}
\description{
Start the EPoSMoL Shiny Webapplication.
}
